Use <Suspense> with async setup for Profile data fetching
