+++ src/ECER.Clients.RegistryPortal/ecer.clients.registryportal.client/src/components/Messages.vue
@@ -1,5 +1,9 @@
 <template>
-  <h1>Messages</h1>
+  <v-row class="ma-4">
+    <v-col cols="12">
+      <h1>Messages</h1>
+    </v-col>
+  </v-row>
 </template>
 
 <script lang="ts">
